+++ Sources/Dependencies/Documentation.docc/Articles/Lifetimes.md
@@ -218,9 +218,10 @@ feature behaves in very specific states. For example, if you wanted to see how y
 when the `fetchUser` endpoint throws an error, you can update the preview like so:
 
 ```swift
-#Preview(traits: .dependencies {
-  $0.apiClient.fetchUser = { _ in throw SomeError() }
-}) {
+#Preview {
+  let _ = prepareDependencies {
+    $0.apiClient.fetchUser = { _ in throw SomeError() }
+  }
   FeatureView(model: FeatureModel())
 }
 ```
