+++ Sources/Dependencies/Documentation.docc/Articles/WhatAreDependencies.md
@@ -99,12 +99,13 @@ final class FeatureModel {
 
 That small change makes this feature much friendlier to Xcode previews and testing.
 
-For previews, you can use the `.dependencies` preview trait to override the
+For previews, you can use ``prepareDependencies(_:)`` to override the
 ``DependencyValues/continuousClock`` dependency to be an "immediate" clock, which is a clock that
 does not actually sleep for any amount of time:
 
 ```swift
-#Preview(traits: .dependencies { $0.continuousClock = ImmediateClock() }) {
+#Preview {
+  let _ = prepareDependencies { $0.continuousClock = ImmediateClock() }
   FeatureView(model: FeatureModel())
 }
 ```
